requirements.txt
fastapi
aiohttp
python-dotenv
requests
